requirements.txt
aiohttp>=3.9.0
supabase>=2.0.0
python-dotenv>=1.0.0
